Fork Recovery 1:#
release 2.9.10 of the main fork head, had a bug, db4ple supplyed a bugfix for this, which i applied succesfully, but then they release 2.9.11 it through me out of sync and i could not merge, I removed the patch from my files. But it still would not merge, removed my github and resynced, ment to apply my changes back to the fork but my work ended up getting over written. This src base has has been recovered from 2.9.4 but i did loose several days of my work. There is now a BACKUP fork which i will be using periodically and also backups of backups locally for redundancy. I do not want to face this situation again in the future.